1. VM on fragments always differ 2. path param on retrofit is Int 3. check build requests

+++ app/src/main/java/com/example/thindie/astonrickandmorty/data/episodes/EpisodesRepositoryImpl.kt
@@ -40,7 +40,7 @@ class EpisodesRepositoryImpl @Inject constructor(
 
     override suspend fun getConcrete(concretes: List<String>): Result<List<EpisodeDomain>> {
         return outSourceLogic.getConcrete(concretes, episodesDtoToDomain) { path ->
-            api.getMultiply(path)
+            api.getMultiply(path.toInt())
         }
             .onFailure {
                 outSourceLogic.onFailedFetchConcrete(episodesDbModelToDomain)
